% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_assays_for_disease_li}
\alias{list_assays_for_disease_li}
\title{List Available Assays for a Disease}
\usage{
list_assays_for_disease_li(diseaseCode)
}
\arguments{
\item{diseaseCode}{TCGA disease code (e.g., "BRCA", "COAD")}
}
\value{
List with disease code, version, and available assays
}
\description{
List Available Assays for a Disease
}
\note{
we hard code version to 2.1.1
}
