% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_cancer_types_df}
\alias{list_cancer_types_df}
\title{List Available TCGA Cancer Types}
\usage{
list_cancer_types_df()
}
\value{
A data.frame with disease codes, names, and subtype availability
}
\description{
List Available TCGA Cancer Types
}
