% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_cancer_types_li}
\alias{list_cancer_types_li}
\title{Produce list of code-name pairs for TCGA Cancer Types that have available assays according to TCGAutils}
\usage{
list_cancer_types_li()
}
\value{
A data.frame with disease codes, names, and subtype availability
}
\description{
Produce list of code-name pairs for TCGA Cancer Types that have available assays according to TCGAutils
}
